refactor: move out MSAFunctionMerger to header
