+++ lib/Transforms/IPO/MultipleSequenceAlignment.cpp
@@ -39,29 +39,6 @@
 
 namespace llvm {
 
-struct MSAFunctionMergeResult {};
-
-/// \brief This pass merges multiple functions into a single function by
-/// multiple sequence alignment algorithm.
-class MSAFunctionMerger {
-  FunctionMerger PairMerger;
-  ScoringSystem Scoring;
-  Module *M;
-
-public:
-  MSAFunctionMerger(Module *M)
-      : PairMerger(M), Scoring(/*Gap*/ 1, /*Match*/ 0, /*Mismatch*/ 1), M(M) {}
-
-  FunctionMerger &getPairMerger() { return PairMerger; }
-
-  void align(const SmallVectorImpl<SmallVectorImpl<Value *> *> &InstrVecRefList,
-             std::vector<MSAAlignmentEntry> &Alignment);
-  MSAFunctionMergeResult merge(ArrayRef<Function *> Functions);
-  void merge(ArrayRef<Function *> Functions,
-             const SmallVectorImpl<SmallVectorImpl<Value *> *> &InstrVecRefList,
-             const std::vector<MSAAlignmentEntry> &Alignment);
-};
-
 class MSAGenFunctionBody;
 
 class MSAGenFunction {
